move validation unit tests to a different file
