+++ python/ribasim/ribasim/model.py
@@ -59,7 +59,7 @@ from ribasim.utils import (
     _node_lookup_numpy,
     _time_in_ns,
 )
-from ribasim.validation import flow_edge_amount
+from ribasim.validation import control_edge_amount, flow_edge_amount
 
 try:
     import xugrid
@@ -302,28 +302,33 @@ class Model(FileModel):
         )
         df_graph = df_graph.rename(columns={"node_type": "to_node_type"})
 
-        if not self._check_neighbors(df_graph, flow_edge_amount, df_node["node_type"]):
-            raise ValueError("Minimum inneighbor or outneighbor unsatisfied")
+        if not self._check_neighbors(
+            df_graph, flow_edge_amount, "flow", df_node["node_type"]
+        ):
+            raise ValueError("Minimum flow inneighbor or outneighbor unsatisfied")
+        if not self._check_neighbors(
+            df_graph, control_edge_amount, "control", df_node["node_type"]
+        ):
+            raise ValueError("Minimum control inneighbor or outneighbor unsatisfied")
 
     def _check_neighbors(
         self,
         df_graph: pd.DataFrame,
-        flow_edge_amount: dict[str, list[int]],
+        edge_amount: dict[str, list[int]],
+        edge_type: str,
         nodes,
     ) -> bool:
         is_valid = True
-        # Count flow edge neighbor
-        df_graph_flow = df_graph.loc[df_graph["edge_type"] == "flow"]
+        # Count edge neighbor
+        df_graph = df_graph.loc[df_graph["edge_type"] == edge_type]
 
         # check from node's neighbor
         from_node_count = (
-            df_graph_flow.groupby("from_node_id")
-            .size()
-            .reset_index(name="from_node_count")  # type: ignore
+            df_graph.groupby("from_node_id").size().reset_index(name="from_node_count")  # type: ignore
         )
 
         df_result = (
-            df_graph_flow[["from_node_id", "from_node_type"]]
+            df_graph[["from_node_id", "from_node_type"]]
             .drop_duplicates()
             .merge(from_node_count, on="from_node_id", how="left")
         )
@@ -342,19 +347,19 @@ class Model(FileModel):
         for _, row in df_result.iterrows():
             # from node's outneighbor
             try:
-                if row["from_node_count"] < flow_edge_amount[row["from_node_type"]][2]:
+                if row["from_node_count"] < edge_amount[row["from_node_type"]][2]:
                     is_valid = False
                     raise ValueError(
-                        f"Node {row['from_node_id']} must have at least {flow_edge_amount[row['from_node_type']][2]} outneighbor(s) (got {row['from_node_count']})"
+                        f"Node {row['from_node_id']} must have at least {edge_amount[row['from_node_type']][2]} outneighbor(s) (got {row['from_node_count']})"
                     )
             except ValueError as e:
                 logging.error(e)
         # check to node's neighbor
         to_node_count = (
-            df_graph_flow.groupby("to_node_id").size().reset_index(name="to_node_count")  # type: ignore
+            df_graph.groupby("to_node_id").size().reset_index(name="to_node_count")  # type: ignore
         )
         df_result = (
-            df_graph_flow[["to_node_id", "to_node_type"]]
+            df_graph[["to_node_id", "to_node_type"]]
             .drop_duplicates()
             .merge(to_node_count, on="to_node_id", how="left")
         )
@@ -372,10 +377,10 @@ class Model(FileModel):
 
         for _, row in df_result.iterrows():
             try:
-                if row["to_node_count"] < flow_edge_amount[row["to_node_type"]][0]:
+                if row["to_node_count"] < edge_amount[row["to_node_type"]][0]:
                     is_valid = False
                     raise ValueError(
-                        f"Node {row['to_node_id']} must have at least {flow_edge_amount[row['to_node_type']][0]} inneighbor(s) (got {row['to_node_count']})"
+                        f"Node {row['to_node_id']} must have at least {edge_amount[row['to_node_type']][0]} inneighbor(s) (got {row['to_node_count']})"
                     )
             except ValueError as e:
                 logging.error(e)
